--- cronv_iter.go
@@ -1,55 +1,0 @@
-package cronv
-
-import (
-	"github.com/gorhill/cronexpr"
-	"time"
-)
-
-type Cronv struct {
-	Crontab         *Crontab
-	expr            *cronexpr.Expression
-	startTime       time.Time
-	durationMinutes float64
-}
-
-func NewCronv(line string, startTime time.Time, durationMinutes float64) (*Cronv, error) {
-	crontab, err := ParseCrontab(line)
-	if err != nil {
-		return nil, err
-	}
-
-	expr, err := cronexpr.Parse(crontab.Schedule.ToCrontab())
-	if err != nil {
-		return nil, err
-	}
-
-	cronv := &Cronv{
-		Crontab:         crontab,
-		expr:            expr,
-		startTime:       startTime,
-		durationMinutes: durationMinutes,
-	}
-	return cronv, nil
-}
-
-type Exec struct {
-	Start time.Time
-	End   time.Time
-}
-
-func (self *Cronv) Iter() <-chan *Exec {
-	ch := make(chan *Exec)
-	eneTime := self.startTime.Add(time.Duration(self.durationMinutes) * time.Minute)
-	next := self.expr.Next(self.startTime)
-	go func() {
-		for next.Equal(eneTime) || eneTime.After(next) {
-			ch <- &Exec{
-				Start: next,
-				End:   next.Add(time.Duration(1) * time.Minute),
-			}
-			next = self.expr.Next(next)
-		}
-		close(ch)
-	}()
-	return ch
-}
